requirements.txt
requests
torch
transformers>=4.40.0
sentence-transformers
faiss-cpu
accelerate
bitsandbytes
pandas
numpy
pickle-mixin 